Mundo: Bean changed

+++ app/src/main/java/com/wallet/crypto/trustapp/ui/widget/holder/WalletHolder.java
@@ -1,13 +1,16 @@
 package com.wallet.crypto.trustapp.ui.widget.holder;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+//import android.support.annotation.NonNull;
+//import android.support.annotation.Nullable;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.RadioButton;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.wallet.crypto.trustapp.R;
 import com.wallet.crypto.trustapp.entity.Wallet;
@@ -73,23 +76,19 @@ public class WalletHolder extends BinderViewHolder<Wallet> implements View.OnCli
 
 	@Override
 	public void onClick(View view) {
-		switch (view.getId()) {
-            case R.id.address:
-			case R.id.default_action: {
-				if (onSetWalletDefaultListener != null) {
-					onSetWalletDefaultListener.onSetDefault(wallet);
-				}
-			} break;
-			case R.id.delete_action: {
-				if (onWalletDeleteListener != null) {
-					onWalletDeleteListener.onDelete(wallet);
-				}
-			} break;
-            case R.id.export_action: {
-                if (onExportWalletListener != null) {
-                    onExportWalletListener.onExport(wallet);
-                }
-            } break;
+		int id = view.getId();
+		if (id == R.id.address || id == R.id.default_action) {
+			if (onSetWalletDefaultListener != null) {
+				onSetWalletDefaultListener.onSetDefault(wallet);
+			}
+		} else if (id == R.id.delete_action) {
+			if (onWalletDeleteListener != null) {
+				onWalletDeleteListener.onDelete(wallet);
+			}
+		} else if (id == R.id.export_action) {
+			if (onExportWalletListener != null) {
+				onExportWalletListener.onExport(wallet);
+			}
 		}
 	}
 }
